# services/orca-core/src/main/resources/openapi/orca-core.yaml
openapi: 3.0.3
...
info:
  title: orca-core
  description: |
    The world as configured (§C1) — tenancy, sites and lanes, users, devices, design artifacts, custom entities, licence verification.
  version: 0.0.1
servers:
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /api/v1/break-templates:
    get:
      tags: [breakTemplates]
      operationId: listBreakTemplates
      summary: The break template catalog, timings included
      responses:
        "200":
          description: The templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakTemplatesEnvelope"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
    post:
      tags: [breakTemplates]
      operationId: createBreakTemplate
      summary: Create a break template with its timings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBreakTemplateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakTemplateEnvelope"
        "400":
          description: The request repeats a break start time
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "409":
          description: An active template already carries that name
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
  /api/v1/break-templates/{breakTemplateExternalId}:
    patch:
      tags: [breakTemplates]
      operationId: updateBreakTemplate
      summary: Update or retire a break template
      description: |
        Absent fields are unchanged; `timings`, when present, REPLACES the
        timing set. A template still referenced by an active team cannot be
        retired.
      parameters:
        - name: breakTemplateExternalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBreakTemplateRequest"
      responses:
        "200":
          description: The template after the change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakTemplateEnvelope"
        "400":
          description: The request repeats a break start time
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "401":
          description: No credential, or one that did not verify
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "404":
          description: No such template
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        "409":
          description: Name in use, or the template is still referenced by an active team
          content:
            application/json:
              schema:
                $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
...
components:
  schemas:
...
    LocalTimeString:
      type: string
      description: A wall-clock time of day, HH:mm or HH:mm:ss.
      pattern: "^([01][0-9]|2[0-3]):[0-5][0-9](:[0-5][0-9])?$"
      example: "06:00"
...
    BreakTimingItem:
      type: object
      required: [startTime, durationMinutes]
      properties:
        startTime:
          $ref: "#/components/schemas/LocalTimeString"
        durationMinutes:
          type: integer
          format: int32
          minimum: 1
    BreakTemplateSummary:
      type: object
      required: [externalId, name, timings, retired]
      properties:
        externalId:
          type: string
        name:
          type: string
        description:
          type: string
        timings:
          type: array
          items:
            $ref: "#/components/schemas/BreakTimingItem"
        retired:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateBreakTemplateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        timings:
          type: array
          items:
            $ref: "#/components/schemas/BreakTimingItem"
    UpdateBreakTemplateRequest:
      type: object
      description: Absent fields are unchanged; `timings` replaces the set when present.
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        timings:
          type: array
          items:
            $ref: "#/components/schemas/BreakTimingItem"
        retired:
          type: boolean
    BreakTemplateEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/BreakTemplateSummary"
    BreakTemplatesEnvelope:
      allOf:
        - $ref: "../../../../../../platform/web/src/main/resources/openapi/_shared.yaml#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/BreakTemplateSummary"
...